+++ packages/appkit/src/networks/ethereum/index.ts
@@ -1,0 +1,2 @@
+export * from './etherTestnet.js'
+export * from './etherMainnet.js'