+++ packages/appkit/src/networks/index.ts
@@ -7,6 +7,7 @@ export * from './bitcoin.js'
 export * from './cross/index.js'
 export * from './bsc/index.js'
 export * from './kaia/index.js'
+export * from './ethereum/index.js'
 // -- Utils ------------------------------------------------------------------
 export * from './utils.js'