+++ packages/sdk/src/react.ts
@@ -1,4 +1,5 @@
 import { EthersAdapter } from '@to-nexus/appkit-adapter-ethers'
+import type { AppKitNetwork } from '@to-nexus/appkit-common'
 import {
   AccountController,
   ApiController,
@@ -12,6 +13,8 @@ import {
   bscTestnet,
   crossMainnet,
   crossTestnet,
+  etherMainnet,
+  etherTestnet,
   kaiaMainnet,
   kaiaTestnet,
   networkList
@@ -51,13 +54,7 @@ export type Metadata = {
   icons: string[]
 }
 
-type SupportedNetworks =
-  | typeof crossTestnet
-  | typeof crossMainnet
-  | typeof bscTestnet
-  | typeof bscMainnet
-  | typeof kaiaTestnet
-  | typeof kaiaMainnet
+type SupportedNetworks = AppKitNetwork
 
 const defaultMetadata: Metadata = {
   name: 'Cross SDK',
@@ -153,6 +150,8 @@ export {
   bscTestnet,
   kaiaMainnet,
   kaiaTestnet,
+  etherMainnet,
+  etherTestnet,
   UniversalProvider,
   getUniversalProvider,
   ConstantsUtil