+++ examples/sdk-vanilla/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @to-nexus/sdk-example-html
 
+## 1.12.3
+
+### Patch Changes
+
+- add ethereum and seporia network
+- Updated dependencies [98e5310]
+- Updated dependencies [de4ba85]
+- Updated dependencies [50c2f26]
+- Updated dependencies [eed3d6d]
+- Updated dependencies
+- Updated dependencies [a6db7b1]
+  - @to-nexus/sdk@1.16.0
+
 ## 1.12.2
 
 ### Patch Changes