+++ packages/adapters/wagmi/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @reown/appkit-adapter-wagmi
 
+## 1.16.0
+
+### Minor Changes
+
+- eed3d6d: disable testnet property of crossMainnet
+
+### Patch Changes
+
+- 98e5310: add logs of detected chain id
+- de4ba85: Remove unmatched native balance cache
+- 50c2f26: add logs for chain id storing procedure
+- a6db7b1: auto switch wallet network uses active network, not default network
+- Updated dependencies [98e5310]
+- Updated dependencies [de4ba85]
+- Updated dependencies [50c2f26]
+- Updated dependencies [eed3d6d]
+- Updated dependencies
+- Updated dependencies [a6db7b1]
+  - @to-nexus/appkit-utils@1.16.0
+  - @to-nexus/appkit-scaffold-ui@1.16.0
+  - @to-nexus/appkit@1.16.0
+  - @to-nexus/appkit-common@1.16.0
+  - @to-nexus/appkit-wallet@1.16.0
+  - @to-nexus/appkit-core@1.16.0
+  - @to-nexus/appkit-ui@1.16.0
+
 ## 1.13.0
 
 ### Patch Changes